Resolve issue if option vlaue has already been deleted

+++ wp-content/uploads/civicrm/ext/biaproperty/CRM/Biaproperty/Upgrader.php
@@ -269,7 +269,9 @@ class CRM_Biaproperty_Upgrader extends CRM_Extension_Upgrader_Base {
       $contactCount = \Civi\Api4\Contact::get(FALSE)->addWhere('Business_Category.Child_Class_Unique', '=', $biaHeading)->execute();
       if (count($contactCount) < 1) {
         $optionValue = \Civi\Api4\OptionValue::get(FALSE)->addWhere('value', '=', $biaHeading)->addWhere('option_group_id:name', '=', 'Business_Category_Child_Class_Unique')->execute()->first();
-        \Civi\Api4\OptionValue::delete(FALSE)->addWhere('id', '=', $optionValue['id'])->execute();
+        if (!empty($optionValue)) {
+          \Civi\Api4\OptionValue::delete(FALSE)->addWhere('id', '=', $optionValue['id'])->execute();
+        }
       }
     }
     return TRUE;
